move PartialPod to pkg/chaosctl
